requirements.txt
requests~=2.21.0
PyYAML~=5.4.1
pandas~=1.2.1